backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pydantic
pytest
httpx
pydub
python-multipart
google-genai